pd: reuse common functions
