+++ src/pd/async/client.rs
@@ -12,29 +12,17 @@
 // limitations under the License.
 
 use std::fmt;
-use std::time::Duration;
-use std::thread;
-use std::collections::HashSet;
 
-use grpc;
-use grpc::futures_grpc::GrpcFutureSend;
-use url::Url;
-use rand::{self, Rng};
 use protobuf::RepeatedField;
 use futures::Future;
 
 use kvproto::metapb;
-use kvproto::pdpb::{self, ErrorType};
-use kvproto::pdpb::{GetMembersRequest, GetMembersResponse, Member};
-use kvproto::pdpb_grpc::PDAsync;
-use kvproto::pdpb_grpc::PDAsyncClient;
-
-use util::HandyRwLock;
+use kvproto::pdpb::{self, Member};
+use kvproto::pdpb_grpc::{PDAsync, PDAsyncClient};
 
 use super::super::PdFuture;
 use super::super::{Result, Error, PdClient};
-use super::util::LeaderClient;
-use super::super:: metrics::*;
+use super::util::*;
 
 pub struct RpcClient {
     cluster_id: u64,
@@ -67,170 +55,7 @@ impl RpcClient {
 
     // For tests
     pub fn get_leader(&self) -> Member {
-        self.leader_client.inner.rl().members.get_leader().clone()
-    }
-}
-
-pub fn validate_endpoints(endpoints: &[String]) -> Result<(PDAsyncClient, GetMembersResponse)> {
-    if endpoints.is_empty() {
-        return Err(box_err!("empty PD endpoints"));
-    }
-
-    let len = endpoints.len();
-    let mut endpoints_set = HashSet::with_capacity(len);
-
-    let mut members = None;
-    let mut cluster_id = None;
-    for ep in endpoints {
-        if !endpoints_set.insert(ep) {
-            return Err(box_err!("duplicate PD endpoint {}", ep));
-        }
-
-        let (_, resp) = match connect(ep) {
-            Ok(resp) => resp,
-            // Ignore failed PD node.
-            Err(e) => {
-                error!("PD endpoint {} failed to respond: {:?}", ep, e);
-                continue;
-            }
-        };
-
-        // Check cluster ID.
-        let cid = resp.get_header().get_cluster_id();
-        if let Some(sample) = cluster_id {
-            if sample != cid {
-                return Err(box_err!("PD response cluster_id mismatch, want {}, got {}",
-                                    sample,
-                                    cid));
-            }
-        } else {
-            cluster_id = Some(cid);
-        }
-        // TODO: check all fields later?
-
-        if members.is_none() {
-            members = Some(resp);
-        }
-    }
-
-    match members {
-        Some(members) => {
-            let (client, members) = try!(try_connect_leader(&members));
-            info!("All PD endpoints are consistent: {:?}", endpoints);
-            Ok((client, members))
-        }
-        _ => Err(box_err!("PD cluster failed to respond")),
-    }
-}
-
-fn connect(addr: &str) -> Result<(PDAsyncClient, GetMembersResponse)> {
-    debug!("connect to PD endpoint: {:?}", addr);
-    let ep = box_try!(Url::parse(addr));
-    let host = ep.host_str().unwrap();
-    let port = ep.port().unwrap();
-
-    let mut conf: grpc::client::GrpcClientConf = Default::default();
-    conf.http.no_delay = Some(true);
-
-    // TODO: It seems that `new` always return an Ok(_).
-    PDAsyncClient::new(host, port, false, conf)
-        .and_then(|client| {
-            // try request.
-            match client.GetMembers(GetMembersRequest::new()).wait() {
-                Ok(resp) => Ok((client, resp)),
-                Err(e) => Err(e),
-            }
-        })
-        .map_err(Error::Grpc)
-}
-
-pub fn try_connect_leader(previous: &GetMembersResponse)
-                          -> Result<(PDAsyncClient, GetMembersResponse)> {
-    // Try to connect other members.
-    // Randomize endpoints.
-    let members = previous.get_members();
-    let mut indexes: Vec<usize> = (0..members.len()).collect();
-    rand::thread_rng().shuffle(&mut indexes);
-
-    let mut resp = None;
-    'outer: for i in indexes {
-        for ep in members[i].get_client_urls() {
-            match connect(ep.as_str()) {
-                Ok((_, r)) => {
-                    resp = Some(r);
-                    break 'outer;
-                }
-                Err(e) => {
-                    error!("failed to connect to {}, {:?}", ep, e);
-                    continue;
-                }
-            }
-        }
-    }
-
-    // Then try to connect the PD cluster leader.
-    if let Some(resp) = resp {
-        let leader = resp.get_leader().clone();
-        for ep in leader.get_client_urls() {
-            if let Ok((client, _)) = connect(ep.as_str()) {
-                info!("connect to PD leader {:?}", ep);
-                return Ok((client, resp));
-            }
-        }
-    }
-
-    Err(box_err!("failed to connect to {:?}", members))
-}
-
-const MAX_RETRY_COUNT: usize = 100;
-const RETRY_INTERVAL: u64 = 1;
-
-fn do_request<F, R>(client: &RpcClient, f: F) -> Result<R>
-    where F: Fn(&PDAsyncClient) -> GrpcFutureSend<R>
-{
-    for _ in 0..MAX_RETRY_COUNT {
-        let r = {
-            let timer = PD_SEND_MSG_HISTOGRAM.start_timer();
-            let r = f(&client.leader_client.inner.rl().client).wait();
-            timer.observe_duration();
-            r
-        };
-
-        match r {
-            Ok(r) => {
-                return Ok(r);
-            }
-            Err(e) => {
-                error!("fail to request: {:?}", e);
-                let result = try_connect_leader(&client.leader_client.inner.rl().members);
-                match result {
-                    Ok((cli, mbrs)) => {
-                        let mut inner = client.leader_client.inner.wl();
-                        inner.client = cli;
-                        inner.members = mbrs;
-                    }
-                    Err(e) => {
-                        error!("fail to connect to PD leader {:?}", e);
-                        thread::sleep(Duration::from_secs(RETRY_INTERVAL));
-                    }
-                }
-            }
-        }
-    }
-
-    Err(box_err!("fail to request"))
-}
-
-fn check_resp_header(header: &pdpb::ResponseHeader) -> Result<()> {
-    if !header.has_error() {
-        return Ok(());
-    }
-    // TODO: translate more error types
-    let err = header.get_error();
-    match err.get_field_type() {
-        ErrorType::ALREADY_BOOTSTRAPPED => Err(Error::ClusterBootstrapped(header.get_cluster_id())),
-        ErrorType::NOT_BOOTSTRAPPED => Err(Error::ClusterNotBootstrapped(header.get_cluster_id())),
-        _ => Err(box_err!(err.get_message())),
+        self.leader_client.get_leader()
     }
 }
 
@@ -255,7 +80,9 @@ impl PdClient for RpcClient {
         req.set_store(stores);
         req.set_region(region);
 
-        let resp = try!(do_request(self, |client| client.Bootstrap(req.clone())));
+        let resp = try!(sync_request(&self.leader_client,
+                                     |client| client.Bootstrap(req.clone()),
+                                     LEADER_CHANGE_RETRY));
         try!(check_resp_header(resp.get_header()));
         Ok(())
     }
@@ -264,7 +91,9 @@ impl PdClient for RpcClient {
         let mut req = pdpb::IsBootstrappedRequest::new();
         req.set_header(self.header());
 
-        let resp = try!(do_request(self, |client| client.IsBootstrapped(req.clone())));
+        let resp = try!(sync_request(&self.leader_client,
+                                     |client| client.IsBootstrapped(req.clone()),
+                                     LEADER_CHANGE_RETRY));
         try!(check_resp_header(resp.get_header()));
 
         Ok(resp.get_bootstrapped())
@@ -274,7 +103,9 @@ impl PdClient for RpcClient {
         let mut req = pdpb::AllocIDRequest::new();
         req.set_header(self.header());
 
-        let resp = try!(do_request(self, |client| client.AllocID(req.clone())));
+        let resp = try!(sync_request(&self.leader_client,
+                                     |client| client.AllocID(req.clone()),
+                                     LEADER_CHANGE_RETRY));
         try!(check_resp_header(resp.get_header()));
 
         Ok(resp.get_id())
@@ -285,7 +116,9 @@ impl PdClient for RpcClient {
         req.set_header(self.header());
         req.set_store(store);
 
-        let resp = try!(do_request(self, |client| client.PutStore(req.clone())));
+        let resp = try!(sync_request(&self.leader_client,
+                                     |client| client.PutStore(req.clone()),
+                                     LEADER_CHANGE_RETRY));
         try!(check_resp_header(resp.get_header()));
 
         Ok(())
@@ -296,7 +129,9 @@ impl PdClient for RpcClient {
         req.set_header(self.header());
         req.set_store_id(store_id);
 
-        let mut resp = try!(do_request(self, |client| client.GetStore(req.clone())));
+        let mut resp = try!(sync_request(&self.leader_client,
+                                         |client| client.GetStore(req.clone()),
+                                         LEADER_CHANGE_RETRY));
         try!(check_resp_header(resp.get_header()));
 
         Ok(resp.take_store())
@@ -306,7 +141,9 @@ impl PdClient for RpcClient {
         let mut req = pdpb::GetClusterConfigRequest::new();
         req.set_header(self.header());
 
-        let mut resp = try!(do_request(self, |client| client.GetClusterConfig(req.clone())));
+        let mut resp = try!(sync_request(&self.leader_client,
+                                         |client| client.GetClusterConfig(req.clone()),
+                                         LEADER_CHANGE_RETRY));
         try!(check_resp_header(resp.get_header()));
 
         Ok(resp.take_cluster())
@@ -317,7 +154,9 @@ impl PdClient for RpcClient {
         req.set_header(self.header());
         req.set_region_key(key.to_vec());
 
-        let mut resp = try!(do_request(self, |client| client.GetRegion(req.clone())));
+        let mut resp = try!(sync_request(&self.leader_client,
+                                         |client| client.GetRegion(req.clone()),
+                                         LEADER_CHANGE_RETRY));
         try!(check_resp_header(resp.get_header()));
 
         Ok(resp.take_region())
